removing unused component and moving form failure dialog closer to odkwebform component
